refactor(ai-engine): resolve #1742 — split bedrock_builder.py (41K) into subpackage
