+++ ai-engine/tests/test_bedrock_builder_mvp.py
@@ -243,7 +243,7 @@ class TestBedrockBuilderMVP:
         rp_path.mkdir()
 
         # Mock Pillow Image processing since we have fake PNG data
-        with patch("agents.bedrock_builder.Image") as mock_image:
+        with patch("agents.bedrock_builder.agent.Image") as mock_image:
             mock_img = Mock()
             mock_img.size = (32, 32)
             mock_img.convert.return_value = mock_img
@@ -360,7 +360,7 @@ class TestBedrockBuilderMVP:
         assert result["success"] is False
         assert len(result["errors"]) > 0
 
-    @patch("agents.bedrock_builder.logger")
+    @patch("agents.bedrock_builder.agent.logger")
     def test_logging_behavior(self, mock_logger, builder, test_jar_with_texture, output_dir):
         """Test that appropriate logging occurs during build."""
         builder.build_block_addon_mvp(
